third_party/WebKit/LayoutTests/fast/css/invalidation/hover-first-letter-sibling-expected.txt

Sibling invalidation with :hover and ::first-letter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target).backgroundColor is transparent
PASS window.eventSender is defined.
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(target).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

This text, and the orange square, should have a green background when the orange square is hovered.
